Terrain Grid border warping prep
Separated out the if/else in InitTerrainTypeGrid. Now the horizontal borders are red and the vertical lines are yellow, to help later distinguish the application of perlin noise. In addition, the grid no longer has red/yellow border lines within a distance which could cause an IndexOutOfBoundsException during the later perlin noise warping.