terrainTypeTexture /w Proper weighting
Added some code to correct the coloring of the input map. Our RGBA values are stored currently in terrainTypeTexture in the pixels. The terrainMiniMap displays our internal colors representing our terrain types. A preliminary form of terrain blending has been implemented, and partially works, but needs a rework due to it's only partially correct result and the very slow runtime; the slow runtime is to be expected as it is a fairly naive/brute force implementation. Currently, the for loop for calling the blending function in InitTerrainTypeGrid is set to not execute.